requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
email-validator>=2.1,<3.0

# Database (MVP foundation)
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0

# Auth / security
passlib[bcrypt]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0
